Move Converter and ValueContext to com.vaadin.data
Part of vaadin/framework8-issues#110
