+++ server/src/main/java/com/vaadin/data/ValueContext.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.data.converter;
+package com.vaadin.data;
 
 import java.io.Serializable;
 import java.util.Locale;
