+++ server/src/main/java/com/vaadin/data/converter/AbstractStringToNumberConverter.java
@@ -20,7 +20,9 @@ import java.text.NumberFormat;
 import java.text.ParsePosition;
 import java.util.Locale;
 
+import com.vaadin.data.Converter;
 import com.vaadin.data.Result;
+import com.vaadin.data.ValueContext;
 
 /**
  * A converter that converts from the number type T to {@link String} and back.
