+++ server/src/main/java/com/vaadin/data/converter/StringToBigDecimalConverter.java
@@ -21,6 +21,7 @@ import java.text.NumberFormat;
 import java.util.Locale;
 
 import com.vaadin.data.Result;
+import com.vaadin.data.ValueContext;
 
 /**
  * A converter that converts from {@link String} to {@link BigDecimal} and back.
