+++ server/src/main/java/com/vaadin/data/converter/StringToBigIntegerConverter.java
@@ -22,6 +22,7 @@ import java.text.NumberFormat;
 import java.util.Locale;
 
 import com.vaadin.data.Result;
+import com.vaadin.data.ValueContext;
 
 /**
  * A converter that converts from {@link String} to {@link BigInteger} and back.
