+++ server/src/main/java/com/vaadin/data/converter/StringToIntegerConverter.java
@@ -20,6 +20,7 @@ import java.text.NumberFormat;
 import java.util.Locale;
 
 import com.vaadin.data.Result;
+import com.vaadin.data.ValueContext;
 
 /**
  * A converter that converts from {@link String} to {@link Integer} and back.
