+++ server/src/main/java/com/vaadin/ui/declarative/converters/DesignObjectConverter.java
@@ -15,9 +15,9 @@
  */
 package com.vaadin.ui.declarative.converters;
 
+import com.vaadin.data.Converter;
 import com.vaadin.data.Result;
-import com.vaadin.data.converter.Converter;
-import com.vaadin.data.converter.ValueContext;
+import com.vaadin.data.ValueContext;
 import com.vaadin.ui.declarative.DesignAttributeHandler;
 
 /**
