+++ server/src/main/java/com/vaadin/ui/declarative/converters/DesignResourceConverter.java
@@ -21,9 +21,9 @@ import java.util.HashMap;
 import java.util.Locale;
 import java.util.Map;
 
+import com.vaadin.data.Converter;
 import com.vaadin.data.Result;
-import com.vaadin.data.converter.Converter;
-import com.vaadin.data.converter.ValueContext;
+import com.vaadin.data.ValueContext;
 import com.vaadin.server.ExternalResource;
 import com.vaadin.server.FileResource;
 import com.vaadin.server.FontAwesome;
