+++ server/src/main/java/com/vaadin/ui/declarative/converters/DesignTimeZoneConverter.java
@@ -17,9 +17,9 @@ package com.vaadin.ui.declarative.converters;
 
 import java.util.TimeZone;
 
+import com.vaadin.data.Converter;
 import com.vaadin.data.Result;
-import com.vaadin.data.converter.Converter;
-import com.vaadin.data.converter.ValueContext;
+import com.vaadin.data.ValueContext;
 import com.vaadin.ui.declarative.DesignAttributeHandler;
 
 /**
