+++ server/src/test/java/com/vaadin/data/BinderBookOfVaadinTest.java
@@ -28,9 +28,7 @@ import org.junit.Test;
 import com.vaadin.data.Binder.Binding;
 import com.vaadin.data.Binder.BindingBuilder;
 import com.vaadin.data.BindingValidationStatus.Status;
-import com.vaadin.data.converter.Converter;
 import com.vaadin.data.converter.StringToIntegerConverter;
-import com.vaadin.data.converter.ValueContext;
 import com.vaadin.data.validator.EmailValidator;
 import com.vaadin.data.validator.StringLengthValidator;
 import com.vaadin.server.AbstractErrorMessage;
