+++ server/src/test/java/com/vaadin/data/BinderTestBase.java
@@ -17,7 +17,6 @@ package com.vaadin.data;
 
 import org.junit.Before;
 
-import com.vaadin.data.converter.Converter;
 import com.vaadin.ui.TextField;
 
 /**
