+++ server/src/test/java/com/vaadin/tests/data/converter/AbstractConverterTest.java
@@ -3,9 +3,9 @@ package com.vaadin.tests.data.converter;
 import org.junit.Assert;
 import org.junit.Test;
 
+import com.vaadin.data.Converter;
 import com.vaadin.data.Result;
-import com.vaadin.data.converter.Converter;
-import com.vaadin.data.converter.ValueContext;
+import com.vaadin.data.ValueContext;
 
 public abstract class AbstractConverterTest {
