+++ server/src/test/java/com/vaadin/tests/data/converter/StringToBooleanConverterTest.java
@@ -7,8 +7,8 @@ import java.util.Locale;
 import org.junit.Assert;
 import org.junit.Test;
 
+import com.vaadin.data.ValueContext;
 import com.vaadin.data.converter.StringToBooleanConverter;
-import com.vaadin.data.converter.ValueContext;
 
 public class StringToBooleanConverterTest extends AbstractStringConverterTest {
