+++ server/src/test/java/com/vaadin/tests/data/converter/StringToDateConverterTest.java
@@ -5,8 +5,8 @@ import java.util.Locale;
 
 import org.junit.Test;
 
+import com.vaadin.data.ValueContext;
 import com.vaadin.data.converter.StringToDateConverter;
-import com.vaadin.data.converter.ValueContext;
 
 public class StringToDateConverterTest extends AbstractConverterTest {
