+++ server/src/test/java/com/vaadin/tests/data/converter/StringToFloatConverterTest.java
@@ -2,8 +2,8 @@ package com.vaadin.tests.data.converter;
 
 import org.junit.Test;
 
+import com.vaadin.data.ValueContext;
 import com.vaadin.data.converter.StringToFloatConverter;
-import com.vaadin.data.converter.ValueContext;
 
 public class StringToFloatConverterTest extends AbstractStringConverterTest {
